+++ src/lib.rs
@@ -21,6 +21,7 @@ pub mod config;
 pub(crate) mod dryrun;
 pub mod install;
 pub mod output;
+pub mod status;
 pub mod sync;
 
 mod app;